add bootstrap script to download small LFS files efficiently
The `create-bootstrap` script searches a repository for smallish LFS
files, combines them into larger LFS files, and adds them to a new
orphan branch called `bootstrap`. In addition, the script adds a `boot`
script to the orphan branch which splits the larger LFS files up, again.

In order to leverage the Git LFS pack files, the Git user needs to get
the `bootstrap` branch and run the `boot` script.